Change size and dimensions to use floats
The size and dimension values across multiple classes have been changed from integers to floats to increase precision. This includes updates to the metric size conversion methods and corresponding tests. These changes should improve the accuracy of size calculations and displays by allowing fractional values.

+++ src/main/kotlin/es/valhalla/web/fermi/engine/component/boxmodel/Border.kt
@@ -19,7 +19,7 @@ data class Border(
 
 data class BorderLine(
 	val line: Line,
-	val size: Int
+	val size: Float
 )
 
 enum class BorderType {
